in process of refactor, datetime conversion problem
